Move APIs into .src file
