+++ .src/ComputeEngineAPI3.java
@@ -1,0 +1,6 @@
+import java.math.BigInteger;
+public interface ComputeEngineAPI3 {
+
+  BigInteger computation(Integer i);
+  
+}
