--- ComputeEngineAPI3.java
@@ -1,6 +1,0 @@
-import java.math.BigInteger;
-public interface ComputeEngineAPI3 {
-
-  BigInteger computation(Integer i);
-  
-}
